+++ src/app/tools/tempus/layout.js
@@ -1,0 +1,25 @@
+import { meta } from "./meta";
+import { generatePreviewMetadata, getFullTitle } from "@/lib/utils";
+
+// Metadata
+export const metadata = {
+	title: meta.title,
+	description: meta.description,
+	keywords: meta.keywords,
+	...generatePreviewMetadata({
+		title: getFullTitle(meta.title),
+		description: meta.description,
+		url: meta.url,
+	}),
+	robots: {
+		index: true,
+		follow: true,
+		nocache: false,
+	},
+};
+
+
+
+export default function TempusToolLayout({ children }) {
+	return children;
+}
