+++ src/app/tools/tempus/meta.ts
@@ -1,0 +1,11 @@
+import type { Meta } from "@/types/meta";
+
+export const meta: Meta = {
+	title: "Tempus",
+	description: "跟著音樂點擊節奏，立即取得並複製精準 BPM，輕鬆對拍。節拍器功能即將登場。",
+	url: "/tools/tempus",
+	keywords: [
+		// Main Keywords
+		"bpm",
+	],
+};
