+++ src/components/main/tools/tempus/tap-area.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useBPMStore } from "@/lib/store/bpm";
+import { useBPMStore } from "@/lib/store/tempus";
 import { cn } from "@/lib/utils";
 
 // Components & UI
